Extract sprite assets (PALETTE, GEORGE frames, and BUG) into a new `georgeSprite` module for better reusability
