+++ src/apps/GeorgeRunnerGameApp.tsx
@@ -1,6 +1,7 @@
 "use client"
 
 import {useRef, useEffect, useState, useCallback} from "react"
+import {PALETTE, GEORGE_RUN_A, GEORGE_RUN_B, GEORGE_JUMP, BUG, type PixelGrid} from "@/os/georgeSprite"
 
 const GAME_W = 600
 const GAME_H = 150
@@ -8,7 +9,7 @@ const GROUND_Y = 130
 
 const HISCORE_KEY = "georgeos:runner:highscore"
 
-type Phase = "ready" | "running" | "gameover"
+type Phase = "running" | "gameover"
 
 interface Obstacle {
   x: number
@@ -57,65 +58,6 @@ function createState(): GameState {
     runFrame: 0,
   }
 }
-
-type PixelGrid = string[]
-
-export const PALETTE: Record<string, string> = {
-  G: "#8b5cf6", // George body (violet)
-  D: "#6d28d9", // George dark accent
-  S: "#f4f6fb", // George highlight / face
-  E: "#1a2233", // eyes / outline
-  B: "#f4796b", // bug body (coral)
-  L: "#1a2233", // bug legs / eyes
-}
-
-// 8 wide x 9 tall — a compact pixel figure (running frame A)
-export const GEORGE_RUN_A: PixelGrid = [
-  "  GGGG  ",
-  " GGGGGG ",
-  " GSSGSG ",
-  " GSEGEG ",
-  " GGGGGG ",
-  "  GGGG  ",
-  " DD  DD ",
-  " D    D ",
-  "DD    DD",
-]
-
-// running frame B — legs swapped
-const GEORGE_RUN_B: PixelGrid = [
-  "  GGGG  ",
-  " GGGGGG ",
-  " GSSGSG ",
-  " GSEGEG ",
-  " GGGGGG ",
-  "  GGGG  ",
-  " DD  DD ",
-  "  D  D  ",
-  "  DD DD ",
-]
-
-// jump pose — legs tucked
-const GEORGE_JUMP: PixelGrid = [
-  "  GGGG  ",
-  " GGGGGG ",
-  " GSSGSG ",
-  " GSEGEG ",
-  " GGGGGG ",
-  "  GGGG  ",
-  " D    D ",
-  "DD    DD",
-  "        ",
-]
-
-// 7 wide x 5 tall bug
-const BUG: PixelGrid = [
-  "L     L",
-  " BBBBB ",
-  "BBLBLBB",
-  " BBBBB ",
-  "L  L  L",
-]
 
 const PLAYER_PX = 4
 const BUG_PX = 3
